requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.0.1
discord.py==1.7.3
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
mysql-connector-python==8.0.32
protobuf==4.24.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
zope.interface==6.0
